skills/docs-hub/requirements-build.txt
pathspec>=0.12,<0.13
PyYAML>=6,<7
markdown-it-py>=3,<4
